episode numbering fixed
fixed the episode numbering to search for all events on events that were matched by round (if they were matched by date we already have that data) and use the strtimestamp to sort and then get an order number.  add that order number to end of episode number to allow events from the same day to be labelled differently and to be in proper order.