custom image deletion system
uses a json file to record the custom images that the agent has created and deletes them during subsequent updates if the videos have been removed.

It also refreshes metadata if it's an individual event (when the refresh metadata button is pressed on a video). At least this way if you want to refresh one event you can without it skipping because the thumbnail exists. (Apparently there's no way for Plex agents to discern if it a refresh metadata) So this is a workaround to bypass my previous fix to stop processing files based on if they already have art applied.